BAL-90-GP-Project Clean, Fix keyboard on signin
